% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.sentiments.R
\name{extract.sentiments}
\alias{extract.sentiments}
\title{Sentiment Analysis of Tweets}
\usage{
\method{extract}{sentiments}(df)
}
\arguments{
\item{df}{A dataset containing a column named 'text'}
}
\value{
Ggplot Graph
}
\description{
Sentiment Analysis of Tweets
}
\examples{
extract.sentiments(tweets)

}
