% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sentiments.R
\name{extract_sentiments}
\alias{extract_sentiments}
\title{Sentiment Analysis of Tweets}
\usage{
extract_sentiments(df)
}
\arguments{
\item{df}{A dataset containing a column named 'text'}
}
\value{
Ggplot Graph
}
\description{
Sentiment Analysis of Tweets
}
\examples{
tweets <- Rtwitter::tw
extract_sentiments(tweets)

}
