% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most.active.R
\name{most.active}
\alias{most.active}
\title{See Most Active Accounts that Post Most Tweets}
\usage{
most.active(df, k = 15)
}
\arguments{
\item{df}{A dataset containing a column named 'user_username'}
}
\value{
Wordcloud
}
\description{
See Most Active Accounts that Post Most Tweets
}
\examples{
tweets <- Rtwitter::tw
most.active(tweets)

}
