% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most.emo.R
\name{most.emo}
\alias{most.emo}
\title{See Most Used Emojis in Tweets}
\usage{
most.emo(df)
}
\arguments{
\item{df}{A dataset containing a column named 'text'}
}
\value{
Table
}
\description{
See Most Used Emojis in Tweets
}
\examples{
tweets <- Rtwitter::tw
most.emo(tweets)

}
