% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most.freq.words.R
\name{most.freq.words}
\alias{most.freq.words}
\title{See Most Frequently Used Words in Tweets}
\usage{
most.freq.words(df)
}
\arguments{
\item{df}{A dataset containing a column named 'text'}
}
\value{
Ggplot Graph
}
\description{
See Most Frequently Used Words in Tweets
}
\examples{
most.freq.words(tweets)

}
