% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most.liked.R
\name{most.liked}
\alias{most.liked}
\title{See Most Liked Tweets}
\usage{
most.liked(df)
}
\arguments{
\item{df}{A dataset containing a column named 'sourcetweet_type', 'text', and 'like_count'}
}
\value{
Ggplot Graph
}
\description{
See Most Liked Tweets
}
\examples{
tweets <- Rtwitter::tw
most.liked(tweets)

}
