% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most.retweets.R
\name{most.retweets}
\alias{most.retweets}
\title{See Most Retweeted Tweets}
\usage{
most.retweets(df)
}
\arguments{
\item{df}{A dataset containing a column named 'sourcetweet_type', 'text', and 'retweet_count'}
}
\value{
Ggplot Graph
}
\description{
See Most Retweeted Tweets
}
\examples{
most.retweets(tweets)

}
