% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/over.time.R
\name{over.time}
\alias{over.time}
\title{See the FREQUENCY OF TWEETS OVER TIME}
\usage{
over.time(df, timeframe = "months")
}
\arguments{
\item{df}{A dataset containing a column named 'created_at'}

\item{timeframe}{This could be months (default), years, or days.}
}
\value{
Ggplot object
}
\description{
See the FREQUENCY OF TWEETS OVER TIME
}
\examples{
tweets <- Rtwitter::tw
over.time(tweets, timeframe="months")

}
