% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.hashtags.R
\name{top.hashtags}
\alias{top.hashtags}
\title{See Top Hashtags}
\usage{
top.hashtags(df, k = 20)
}
\arguments{
\item{df}{A dataset containing a column named text}

\item{n}{Number of hashtags you want to see}
}
\value{
Table
}
\description{
This function lets you to view the top hashtags.
}
\examples{
top.hashtags(tweets, n=5)

}
