% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweet.lang.R
\name{tweet.lang}
\alias{tweet.lang}
\title{See Most Frequently Used Languages in Tweets}
\usage{
tweet.lang(df)
}
\arguments{
\item{df}{A dataset containing a column named 'lang'}
}
\value{
Plot
}
\description{
See Most Frequently Used Languages in Tweets
}
\examples{
tweets <- Rtwitter::tw
tweet.lang(tweets)

}
