% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.locations.R
\name{user.locations}
\alias{user.locations}
\title{See Most Frequent User Locations}
\usage{
user.locations(df)
}
\arguments{
\item{df}{A dataset containing a column named text}
}
\value{
Graph
}
\description{
See Most Frequent User Locations
}
\examples{
tweets <- Rtwitter::tw
user.locations(tweets)

}
