% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.randomly.R
\name{view.randomly}
\alias{view.randomly}
\title{View Some Random Tweets}
\usage{
view.randomly(df, n = 3)
}
\arguments{
\item{df}{A tweet dataset with a column named text and another column named user_name}

\item{n}{Number of tweets you want to see}
}
\value{
Table
}
\description{
This function lets you to view a random number of tweets in a nice format.
}
\examples{
tweets <- Rtwitter::tw
view.randomly(tweets, n=5)

}
