% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.hashtags.R
\name{wordcloud.hashtags}
\alias{wordcloud.hashtags}
\title{See Top Hashtags in a Wordcloud}
\usage{
wordcloud.hashtags(df)
}
\arguments{
\item{df}{A dataset containing a column named text}
}
\value{
Wordcloud Graph
}
\description{
This function lets you to view the top 20 hashtags via a wordcloud
}
\examples{
tweets <- Rtwitter::tw
wordcloud.hashtags(tweets)

}
