% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.tweets.R
\name{wordcloud.tweets}
\alias{wordcloud.tweets}
\title{See Most Frequently Used Words in a Wordcloud}
\usage{
wordcloud.tweets(df)
}
\arguments{
\item{df}{A dataset containing a column named text}
}
\value{
Wordcloud Graph
}
\description{
See Most Frequently Used Words in a Wordcloud
}
\examples{
wordcloud.tweets(tweets)

}
