% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word.network.R
\name{word.network}
\alias{word.network}
\title{See How Words Assosciate with Each Other}
\usage{
word.network(df)
}
\arguments{
\item{df}{A dataset containing a column named 'text'}
}
\value{
Interactive Plot
}
\description{
See How Words Assosciate with Each Other
}
\examples{
word.network(tweets)

}
